+++ package.json
@@ -4,7 +4,7 @@
   "description": "Creating a reat estate purchasing app",
   "main": "index.js",
   "scripts": {
-    "start": "node App.js",
+    "start": "node src/App.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
